chore: Adds & integrates dependencies.gradle
In order to manage the dependencies of the project I've added a dependencies.gradle file which will contain the info about the dependencies very module of the project wants to use.

By splitting and setting up dependencies here we can easily manage a multi module gradle project as and when it grows. It also allows us to make sure that all the modules in future would run with parity terms of library versions.

However in order to move fast with the assignment, the upcoming dependencies that I need, I'll add straight away to the build.gradle file.